[generated] source: spec3.sdk.yaml@non-master-spec-e27a05e in mickjermsurawong/no-v-change
