+++ src/main/java/com/stripe/model/Card.java
@@ -256,28 +256,26 @@ public class Card extends ApiResource
   }
 
   /**
-   * Updates the metadata, account holder name, and account holder type of a bank account belonging
-   * to a <a href="/docs/connect/custom-accounts">Custom account</a>, and optionally sets it as the
-   * default for its currency. Other bank account details are not editable by design.
+   * If you need to update only some card details, like the billing address or expiration date, you
+   * can do so without having to re-enter the full card details. Stripe also works directly with
+   * card networks so that your customers can <a
+   * href="https://stripe.com/docs/saving-cards#automatic-card-updates">continue using your
+   * service</a> without interruption.
    *
-   * <p>You can re-enable a disabled bank account by performing an update call without providing any
-   * arguments or changes.
-   *
-   * <p>Update a specified source for a given customer.
+   * <p>Updates a specified card for a given customer.
    */
   public Card update(Map<String, Object> params) throws StripeException {
     return update(params, (RequestOptions) null);
   }
 
   /**
-   * Updates the metadata, account holder name, and account holder type of a bank account belonging
-   * to a <a href="/docs/connect/custom-accounts">Custom account</a>, and optionally sets it as the
-   * default for its currency. Other bank account details are not editable by design.
+   * If you need to update only some card details, like the billing address or expiration date, you
+   * can do so without having to re-enter the full card details. Stripe also works directly with
+   * card networks so that your customers can <a
+   * href="https://stripe.com/docs/saving-cards#automatic-card-updates">continue using your
+   * service</a> without interruption.
    *
-   * <p>You can re-enable a disabled bank account by performing an update call without providing any
-   * arguments or changes.
-   *
-   * <p>Update a specified source for a given customer.
+   * <p>Updates a specified card for a given customer.
    */
   public Card update(Map<String, Object> params, RequestOptions options) throws StripeException {
     String url;
