+++ src/main/java/com/stripe/model/Charge.java
@@ -8,6 +8,7 @@ import com.stripe.exception.StripeException;
 import com.stripe.model.radar.Rule;
 import com.stripe.net.ApiResource;
 import com.stripe.net.RequestOptions;
+import java.util.List;
 import java.util.Map;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -45,14 +46,17 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
 
   /**
    * The application fee (if any) for the charge. [See the Connect
-   * documentation](/docs/connect/direct-charges#collecting-fees) for details.
+   * documentation](https://stripe.com/docs/connect/direct-charges#collecting-fees) for details.
    */
   @SerializedName("application_fee")
   @Getter(lombok.AccessLevel.NONE)
   @Setter(lombok.AccessLevel.NONE)
   ExpandableField<ApplicationFee> applicationFee;
 
-  /** The amount of the application fee (if any) for the resulting payment. */
+  /**
+   * The amount of the application fee (if any) for the charge. [See the Connect
+   * documentation](https://stripe.com/docs/connect/direct-charges#collecting-fees) for details.
+   */
   @SerializedName("application_fee_amount")
   Long applicationFeeAmount;
 
@@ -98,8 +102,8 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
   String description;
 
   /**
-   * The account (if any) the charge was made on behalf of, with an automatic transfer. [See the
-   * Connect documentation](/docs/connect/destination-charges) for details.
+   * ID of an existing, connected Stripe account to transfer funds to if `transfer_data` was
+   * specified in the charge request.
    */
   @SerializedName("destination")
   @Getter(lombok.AccessLevel.NONE)
@@ -114,7 +118,7 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
 
   /**
    * Error code explaining reason for charge failure if available (see [the errors
-   * section](/docs/api#errors) for a list of codes).
+   * section](https://stripe.com/docs/api#errors) for a list of codes).
    */
   @SerializedName("failure_code")
   String failureCode;
@@ -162,7 +166,7 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
 
   /**
    * The account (if any) the charge was made on behalf of without triggering an automatic transfer.
-   * See the [Connect documentation](/docs/connect/charges-transfers) for details.
+   * See the [Connect documentation](https://stripe.com/docs/connect/charges-transfers) for details.
    */
   @SerializedName("on_behalf_of")
   @Getter(lombok.AccessLevel.NONE)
@@ -177,7 +181,7 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
 
   /**
    * Details about whether the payment was accepted, and why. See [understanding
-   * declines](/docs/declines) for details.
+   * declines](https://stripe.com/docs/declines) for details.
    */
   @SerializedName("outcome")
   Outcome outcome;
@@ -235,7 +239,8 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
 
   /**
    * The transfer ID which created this charge. Only present if the charge came from another Stripe
-   * account. [See the Connect documentation](/docs/connect/destination-charges) for details.
+   * account. [See the Connect documentation](https://stripe.com/docs/connect/destination-charges)
+   * for details.
    */
   @SerializedName("source_transfer")
   @Getter(lombok.AccessLevel.NONE)
@@ -262,12 +267,18 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
   @Setter(lombok.AccessLevel.NONE)
   ExpandableField<Transfer> transfer;
 
+  /**
+   * An optional dictionary including the account to automatically transfer to as part of a
+   * destination charge. [See the Connect
+   * documentation](https://stripe.com/docs/connect/destination-charges) for details.
+   */
   @SerializedName("transfer_data")
   TransferData transferData;
 
   /**
    * A string that identifies this transaction as part of a group. See the [Connect
-   * documentation](/docs/connect/charges-transfers#grouping-transactions) for details.
+   * documentation](https://stripe.com/docs/connect/charges-transfers#grouping-transactions) for
+   * details.
    */
   @SerializedName("transfer_group")
   String transferGroup;
@@ -752,12 +763,58 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
   @Getter
   @Setter
   @EqualsAndHashCode(callSuper = false)
+  public static class Level3 extends StripeObject {
+    @SerializedName("customer_reference")
+    String customerReference;
+
+    @SerializedName("line_items")
+    List<LineItem> lineItems;
+
+    @SerializedName("merchant_reference")
+    String merchantReference;
+
+    @SerializedName("shipping_address_zip")
+    String shippingAddressZip;
+
+    @SerializedName("shipping_amount")
+    Long shippingAmount;
+
+    @SerializedName("shipping_from_zip")
+    String shippingFromZip;
+
+    @Getter
+    @Setter
+    @EqualsAndHashCode(callSuper = false)
+    public static class LineItem extends StripeObject {
+      @SerializedName("discount_amount")
+      Long discountAmount;
+
+      @SerializedName("product_code")
+      String productCode;
+
+      @SerializedName("product_description")
+      String productDescription;
+
+      @SerializedName("quantity")
+      Long quantity;
+
+      @SerializedName("tax_amount")
+      Long taxAmount;
+
+      @SerializedName("unit_cost")
+      Long unitCost;
+    }
+  }
+
+  @Getter
+  @Setter
+  @EqualsAndHashCode(callSuper = false)
   public static class Outcome extends StripeObject {
     /**
      * Possible values are `approved_by_network`, `declined_by_network`, `not_sent_to_network`, and
      * `reversed_after_approval`. The value `reversed_after_approval` indicates the payment was
-     * [blocked by Stripe](/docs/declines#blocked-payments) after bank authorization, and may
-     * temporarily appear as "pending" on a cardholder's statement.
+     * [blocked by Stripe](https://stripe.com/docs/declines#blocked-payments) after bank
+     * authorization, and may temporarily appear as "pending" on a cardholder's statement.
      */
     @SerializedName("network_status")
     String networkStatus;
@@ -767,7 +824,7 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
      * blocked by Radar's default block rule have the value `highest_risk_level`. Charges placed in
      * review by Radar's default review rule have the value `elevated_risk_level`. Charges
      * authorized, blocked, or placed in review by custom rules have the value `rule`. See
-     * [understanding declines](/docs/declines) for more details.
+     * [understanding declines](https://stripe.com/docs/declines) for more details.
      */
     @SerializedName("reason")
     String reason;
@@ -805,8 +862,8 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
 
     /**
      * Possible values are `authorized`, `manual_review`, `issuer_declined`, `blocked`, and
-     * `invalid`. See [understanding declines](/docs/declines) and [Radar reviews](radar/review) for
-     * details.
+     * `invalid`. See [understanding declines](https://stripe.com/docs/declines) and [Radar
+     * reviews](radar/review) for details.
      */
     @SerializedName("type")
     String type;
@@ -835,8 +892,15 @@ public class Charge extends ApiResource implements BalanceTransactionSource, Met
   @EqualsAndHashCode(callSuper = false)
   public static class TransferData extends StripeObject {
     /**
-     * The account (if any) the charge was made on behalf of, with an automatic transfer. [See the
-     * Connect documentation](/docs/connect/destination-charges) for details.
+     * The amount transferred to the destination account, if specified. By default, the entire
+     * charge amount is transferred to the destination account.
+     */
+    @SerializedName("amount")
+    Long amount;
+
+    /**
+     * ID of an existing, connected Stripe account to transfer funds to if `transfer_data` was
+     * specified in the charge request.
      */
     @SerializedName("destination")
     @Getter(lombok.AccessLevel.NONE)
