+++ src/main/java/com/stripe/model/Customer.java
@@ -205,9 +205,10 @@ public class Customer extends ApiResource implements HasId, MetadataStore<Custom
    * in the future. When you update a customer to a new valid card source by passing the
    * <strong>source</strong> parameter: for each of the customer’s current subscriptions, if the
    * subscription bills automatically and is in the <code>past_due</code> state, then the latest
-   * unclosed invoice for the subscription will be retried. This retry will not count as an
-   * automatic retry, and will not affect the next regularly scheduled payment for the invoice.
-   * Changing the <strong>default_source</strong> for a customer will not trigger this behavior.
+   * open invoice for the subscription with automatic collection enabled will be retried. This retry
+   * will not count as an automatic retry, and will not affect the next regularly scheduled payment
+   * for the invoice. Changing the <strong>default_source</strong> for a customer will not trigger
+   * this behavior.
    *
    * <p>This request accepts mostly the same arguments as the customer creation call.
    */
@@ -222,9 +223,10 @@ public class Customer extends ApiResource implements HasId, MetadataStore<Custom
    * in the future. When you update a customer to a new valid card source by passing the
    * <strong>source</strong> parameter: for each of the customer’s current subscriptions, if the
    * subscription bills automatically and is in the <code>past_due</code> state, then the latest
-   * unclosed invoice for the subscription will be retried. This retry will not count as an
-   * automatic retry, and will not affect the next regularly scheduled payment for the invoice.
-   * Changing the <strong>default_source</strong> for a customer will not trigger this behavior.
+   * open invoice for the subscription with automatic collection enabled will be retried. This retry
+   * will not count as an automatic retry, and will not affect the next regularly scheduled payment
+   * for the invoice. Changing the <strong>default_source</strong> for a customer will not trigger
+   * this behavior.
    *
    * <p>This request accepts mostly the same arguments as the customer creation call.
    */
