--- src/main/java/com/stripe/model/Level3LineItem.java
@@ -1,31 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class Level3LineItem extends StripeObject {
-  @SerializedName("discount_amount")
-  Long discountAmount;
-
-  @SerializedName("product_code")
-  String productCode;
-
-  @SerializedName("product_description")
-  String productDescription;
-
-  @SerializedName("quantity")
-  Long quantity;
-
-  @SerializedName("tax_amount")
-  Long taxAmount;
-
-  @SerializedName("unit_cost")
-  Long unitCost;
-}
