+++ src/main/java/com/stripe/model/Order.java
@@ -120,7 +120,8 @@ public class Order extends ApiResource implements HasId, MetadataStore<Order> {
 
   /**
    * Current order status. One of `created`, `paid`, `canceled`, `fulfilled`, or `returned`. More
-   * details in the [Orders Guide](/docs/orders/guide#understanding-order-statuses).
+   * details in the [Orders
+   * Guide](https://stripe.com/docs/orders/guide#understanding-order-statuses).
    */
   @SerializedName("status")
   String status;
