+++ src/main/java/com/stripe/model/Plan.java
@@ -132,7 +132,7 @@ public class Plan extends ApiResource implements HasId, MetadataStore<Plan> {
 
   /**
    * Default number of trial days when subscribing a customer to this plan using
-   * [`trial_from_plan=true`](/docs/api#create_subscription-trial_from_plan).
+   * [`trial_from_plan=true`](https://stripe.com/docs/api#create_subscription-trial_from_plan).
    */
   @SerializedName("trial_period_days")
   Long trialPeriodDays;
