--- src/main/java/com/stripe/model/SourceTypeAchCreditTransfer.java
@@ -1,40 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeAchCreditTransfer extends StripeObject {
-  @SerializedName("account_number")
-  String accountNumber;
-
-  @SerializedName("bank_name")
-  String bankName;
-
-  @SerializedName("fingerprint")
-  String fingerprint;
-
-  @SerializedName("refund_account_holder_name")
-  String refundAccountHolderName;
-
-  @SerializedName("refund_account_holder_type")
-  String refundAccountHolderType;
-
-  @SerializedName("refund_account_number")
-  String refundAccountNumber;
-
-  @SerializedName("refund_routing_number")
-  String refundRoutingNumber;
-
-  @SerializedName("routing_number")
-  String routingNumber;
-
-  @SerializedName("swift_code")
-  String swiftCode;
-}
