--- src/main/java/com/stripe/model/SourceTypeAchDebit.java
@@ -1,31 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeAchDebit extends StripeObject {
-  @SerializedName("bank_name")
-  String bankName;
-
-  @SerializedName("country")
-  String country;
-
-  @SerializedName("fingerprint")
-  String fingerprint;
-
-  @SerializedName("last4")
-  String last4;
-
-  @SerializedName("routing_number")
-  String routingNumber;
-
-  @SerializedName("type")
-  String type;
-}
