--- src/main/java/com/stripe/model/SourceTypeAlipay.java
@@ -1,22 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeAlipay extends StripeObject {
-  @SerializedName("data_string")
-  String dataString;
-
-  @SerializedName("native_url")
-  String nativeUrl;
-
-  @SerializedName("statement_descriptor")
-  String statementDescriptor;
-}
