--- src/main/java/com/stripe/model/SourceTypeCardPresent.java
@@ -1,82 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeCardPresent extends StripeObject {
-  @SerializedName("application_cryptogram")
-  String applicationCryptogram;
-
-  @SerializedName("application_preferred_name")
-  String applicationPreferredName;
-
-  @SerializedName("authorization_code")
-  String authorizationCode;
-
-  @SerializedName("authorization_response_code")
-  String authorizationResponseCode;
-
-  @SerializedName("brand")
-  String brand;
-
-  @SerializedName("country")
-  String country;
-
-  @SerializedName("cvm_type")
-  String cvmType;
-
-  @SerializedName("data_type")
-  String dataType;
-
-  @SerializedName("dedicated_file_name")
-  String dedicatedFileName;
-
-  @SerializedName("emv_auth_data")
-  String emvAuthData;
-
-  @SerializedName("evidence_customer_signature")
-  String evidenceCustomerSignature;
-
-  @SerializedName("evidence_transaction_certificate")
-  String evidenceTransactionCertificate;
-
-  @SerializedName("exp_month")
-  Long expMonth;
-
-  @SerializedName("exp_year")
-  Long expYear;
-
-  @SerializedName("fingerprint")
-  String fingerprint;
-
-  @SerializedName("funding")
-  String funding;
-
-  @SerializedName("last4")
-  String last4;
-
-  @SerializedName("pos_device_id")
-  String posDeviceId;
-
-  @SerializedName("pos_entry_mode")
-  String posEntryMode;
-
-  @SerializedName("read_method")
-  String readMethod;
-
-  @SerializedName("reader")
-  String reader;
-
-  @SerializedName("terminal_verification_results")
-  String terminalVerificationResults;
-
-  @SerializedName("transaction_status_information")
-  String transactionStatusInformation;
-}
