--- src/main/java/com/stripe/model/SourceTypeEps.java
@@ -1,19 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeEps extends StripeObject {
-  @SerializedName("reference")
-  String reference;
-
-  @SerializedName("statement_descriptor")
-  String statementDescriptor;
-}
