--- src/main/java/com/stripe/model/SourceTypeMultibanco.java
@@ -1,43 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeMultibanco extends StripeObject {
-  @SerializedName("entity")
-  String entity;
-
-  @SerializedName("reference")
-  String reference;
-
-  @SerializedName("refund_account_holder_address_city")
-  String refundAccountHolderAddressCity;
-
-  @SerializedName("refund_account_holder_address_country")
-  String refundAccountHolderAddressCountry;
-
-  @SerializedName("refund_account_holder_address_line1")
-  String refundAccountHolderAddressLine1;
-
-  @SerializedName("refund_account_holder_address_line2")
-  String refundAccountHolderAddressLine2;
-
-  @SerializedName("refund_account_holder_address_postal_code")
-  String refundAccountHolderAddressPostalCode;
-
-  @SerializedName("refund_account_holder_address_state")
-  String refundAccountHolderAddressState;
-
-  @SerializedName("refund_account_holder_name")
-  String refundAccountHolderName;
-
-  @SerializedName("refund_iban")
-  String refundIban;
-}
