--- src/main/java/com/stripe/model/SourceTypeP24.java
@@ -1,16 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeP24 extends StripeObject {
-  @SerializedName("reference")
-  String reference;
-}
