--- src/main/java/com/stripe/model/SourceTypeSepaDebit.java
@@ -1,37 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeSepaDebit extends StripeObject {
-  @SerializedName("bank_code")
-  String bankCode;
-
-  @SerializedName("branch_code")
-  String branchCode;
-
-  @SerializedName("country")
-  String country;
-
-  @SerializedName("fingerprint")
-  String fingerprint;
-
-  @SerializedName("last4")
-  String last4;
-
-  @SerializedName("mandate_reference")
-  String mandateReference;
-
-  @SerializedName("mandate_url")
-  String mandateUrl;
-
-  @SerializedName("skip_validation")
-  Boolean skipValidation;
-}
