--- src/main/java/com/stripe/model/SourceTypeSofort.java
@@ -1,34 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeSofort extends StripeObject {
-  @SerializedName("bank_code")
-  String bankCode;
-
-  @SerializedName("bank_name")
-  String bankName;
-
-  @SerializedName("bic")
-  String bic;
-
-  @SerializedName("country")
-  String country;
-
-  @SerializedName("iban_last4")
-  String ibanLast4;
-
-  @SerializedName("preferred_language")
-  String preferredLanguage;
-
-  @SerializedName("statement_descriptor")
-  String statementDescriptor;
-}
