--- src/main/java/com/stripe/model/SourceTypeWechat.java
@@ -1,25 +1,0 @@
-// Generated by com.stripe.generator.entity.SdkBuilder
-
-package com.stripe.model;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@EqualsAndHashCode(callSuper = false)
-public class SourceTypeWechat extends StripeObject {
-  @SerializedName("native_url")
-  String nativeUrl;
-
-  @SerializedName("prepay_id")
-  String prepayId;
-
-  @SerializedName("qr_code_url")
-  String qrCodeUrl;
-
-  @SerializedName("statement_descriptor")
-  String statementDescriptor;
-}
