+++ src/main/java/com/stripe/model/Transfer.java
@@ -117,7 +117,8 @@ public class Transfer extends ApiResource
 
   /**
    * A string that identifies this transaction as part of a group. See the [Connect
-   * documentation](/docs/connect/charges-transfers#grouping-transactions) for details.
+   * documentation](https://stripe.com/docs/connect/charges-transfers#grouping-transactions) for
+   * details.
    */
   @SerializedName("transfer_group")
   String transferGroup;
