+++ src/main/java/com/stripe/model/issuing/Settlement.java
@@ -22,6 +22,10 @@ public class Settlement extends ApiResource implements HasId, MetadataStore<Sett
   @SerializedName("bin")
   String bin;
 
+  /** The date that the transactions are cleared and posted to user's accounts. */
+  @SerializedName("clearing_date")
+  Long clearingDate;
+
   /** Time at which the object was created. Measured in seconds since the Unix epoch. */
   @SerializedName("created")
   Long created;
@@ -73,10 +77,6 @@ public class Settlement extends ApiResource implements HasId, MetadataStore<Sett
   @SerializedName("object")
   String object;
 
-  /** The date of the settlement event. */
-  @SerializedName("settlement_date")
-  Long settlementDate;
-
   /** One of `international` or `uk_national_net`. */
   @SerializedName("settlement_service")
   String settlementService;
@@ -88,10 +88,6 @@ public class Settlement extends ApiResource implements HasId, MetadataStore<Sett
   /** The total transaction amount reflected in this settlement. */
   @SerializedName("transaction_volume")
   Long transactionVolume;
-
-  /** Whether or not the transactions have been reconciled. */
-  @SerializedName("transactions_reconciled")
-  Boolean transactionsReconciled;
 
   /**
    * Returns a list of Issuing <code>Settlement</code> objects. The objects are sorted in descending
